Live delay toggle refactor
